+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Reverted to binning with default of 10,000 bins for AbbottFirestoneCurve due to large performance bottleneck of
   functional parameter calculations based on whole dataset
 - Added support for multiple return values in the Batch class
+- Surface.align now takes axis argument to specify with which axis to align the texture
 ## v0.3.0:
 - Removed underscore from Surface._data, ._step_x, ._step_y, ._width_um, ._height_um since they are supposed to be
   public attributes